refactor: extract drop zone handling into `useCanvasDropZone` composable
